remove commented code
